requirements.txt
asn1crypto==0.24.0
audioread==2.1.8
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
dominate==2.3.5
enum34==1.1.6
filetype==1.0.5
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
gevent==1.4.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-datastore==1.8.0
google-cloud-storage==1.17.0
google-cloud-translate==1.6.0
google-cloud-vision==0.40.0
google-images-download==2.8.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.25.0
idna==2.8
imageio==2.5.0
ipaddress==1.0.22
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
MarkupSafe==1.1.1
mat4py==0.4.2
matplotlib==2.2.4
money==1.3.0
networkx==2.3
nltk==3.4.5
numba==0.46.0
numpy==1.17.4
opencv-python==4.1.1.26
pandas==0.24.2
parso==0.5.1
pdf2image==1.7.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==6.1.0
pip-upgrade-outdated==1.5
pipenv==2018.11.26
prompt-toolkit==2.0.10
protobuf==3.10.0
py-money==0.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyCurrency-Converter==1.0
pygame==1.9.6
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
pytesseract==0.2.7
python-dateutil==2.8.0
python-oauth2==1.1.1
pytictoc==1.5.0
pytz==2019.3
PyWavelets==1.0.3
requests==2.22.0
resampy==0.2.2
rsa==4.0
save-ipython-variables==0.0.4
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selectivesearch==0.4
selenium==3.141.0
six==1.13.0
sklearn-export==0.0.4
SoundFile==0.10.2
sqlparse==0.3.0
stanford==0.0.0
stop-words==2018.7.23
subprocess32==3.5.4
tesseract==0.1.3
tesserocr==2.4.0
traitlets==4.3.3
urllib3==1.25.7
virtualenv==16.7.2
virtualenv-clone==0.5.3
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.15.5
